indent-common: validateCore(): Generalize having multiple acceptable indents
Until now, validateCore took as separate arguments 1 or 2 acceptable indent
levels.  Generalize it to take a list of 1 to N acceptable indent levels.
This makes the code clearer.

While we're at it, this makes obvious that it's user-hostile to offer a
single fix if there are multiple acceptable indentation levels.  Make it
clear to the user when there are multiple indents that we'd accept.