New: html-indent: Add "smart-tab" indent type
With the preceding commits generalizing the indent computation logic
to work with tab/space pairs, all that is left to do is add a type
that is like "tab", but doesn't trigger the tab-specific logic when
figuring out alignments.

The bigger part of this commit is updating the tests.  When
automatically generating expected messages, it must now generate
messages that include a combination of tabs and spaces.  To do this,
we export indentHumanize() from indent-common.js.