tests: attribute-hyphenation: Fix the ignore-list test
The test introduced in f9e1eee3fe711b9dfedc037fe25aa5570370296e didn't
actually test what it meant to test for.  It attempts to verify that
the ignore list is obeyed; but its subject is an element that is
ignored in its entirety!  The code under test was never triggered!

Fix this by changing <div> for <custom> in the test case, and also add
the same test-case without the `ignore` config, in order to verify
that it actually has something to ignore.