Fix attribute-hyphenation ignore lists (#513)
* tests: attribute-hyphenation: Fix the ignore-list test

The test introduced in f9e1eee3fe711b9dfedc037fe25aa5570370296e didn't
actually test what it meant to test for.  It attempts to verify that
the ignore list is obeyed; but its subject is an element that is
ignored in its entirety!  The code under test was never triggered!

Fix this by changing <div> for <custom> in the test case, and also add
the same test-case without the `ignore` config, in order to verify
that it actually has something to ignore.

* attribute-hyphenation: Work correctly when ignore has != 1 item

Currently, the "ignore" options array only works correctly with exactly 1
item.  Adjust it to work with 0 and with >1 items.

* attribute-hyphenation: Fix spelling mistakes

 - hypen -> hyphen
 - cann't -> can't

* Update test suite for attribute-hyphenation