Add "ignores" list to prop names casing
Using Vue 3, it is recommended to use the "emits" option to define the events emitted by a component. However, in doing so, the events are not available in the "attrs" object anymore and need to be registered as props. Following the pattern of "update:modelValue", it is very common to use ":" inside an event's name, which is currently not possible using the "vue/prop-name-casing" rule. This commit allows for adding an "ignores" option (array of regex patterns) just like the "custom-event-name-casing" rule. 

PS: A default value for this "ignores" option should probably be [".*?:.*?"].